Moved popcorn player tests to correct package
