--- player-popcorn/src/test/java/com/github/yoep/player/popcorn/PopcornPlayerTest.java
@@ -1,42 +1,0 @@
-package com.github.yoep.player.popcorn;
-
-import com.github.yoep.player.popcorn.player.PopcornPlayer;
-import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.extension.ExtendWith;
-import org.mockito.InjectMocks;
-import org.mockito.junit.jupiter.MockitoExtension;
-
-import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.junit.jupiter.api.Assertions.assertFalse;
-
-@ExtendWith(MockitoExtension.class)
-class PopcornPlayerTest {
-    @InjectMocks
-    private PopcornPlayer popcornPlayer;
-
-    @Test
-    void testGetId_whenInvoked_shouldReturnTheExpectedId() {
-        var result = popcornPlayer.getId();
-
-        assertEquals(PopcornPlayer.PLAYER_ID, result);
-    }
-
-    @Test
-    void testGetName_whenInvoked_shouldReturnTheExpectedName() {
-        var result = popcornPlayer.getName();
-
-        assertEquals(PopcornPlayer.PLAYER_NAME, result);
-    }
-
-    @Test
-    void testIsEmbeddedPlaybackSupported_whenInvoked_shouldReturnTrue() {
-        var result = popcornPlayer.isEmbeddedPlaybackSupported();
-
-        assertFalse(result, "Expected the popcorn player to NOT support embedded playback");
-    }
-
-    @Test
-    void testDispose_whenInvoked_shouldDisposeTheVideoPlayers() {
-        popcornPlayer.dispose();
-    }
-}
